Add Postgres, MsSql, and Oracle support to schema to xml generation #85

+++ reladomo/src/main/java/com/gs/fw/common/mithra/databasetype/MsSqlDatabaseType.java
@@ -454,7 +454,7 @@ public class MsSqlDatabaseType extends AbstractDatabaseType
         {
             if (decimal != 0)
             {
-                javaType =  "double";
+                javaType =  "BigDecimal";
             }
             else if (precision <= 8)
             {
